Source local zshrc if it exists
We've come full circle.

Previously, I thought the best approach would be to have .zshrc be
completely portable across machines, and keep *all* machine-local
configuration in .zprofile. This would imply that all the machine-local
configuration would not rely on any interactive-only sorts of config,
since .zprofile is read by non-interactive shells. However, it's useful
to be able to set up interactive features that are only relevant to
particular machines, for example adding named directories with `hash
-d`. Since .zshrc is the *only* appropriate file for such configuration,
we're back to having a 'portable' zshrc that looks for machine-specific
commands as well. It might be that my original approach is the most
reasonable way to do this after all.